moved Stop and Kill to linux/windows specific files && run linux file on everything except windows
