+++ src/factorio_server.go
@@ -11,12 +11,9 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
+	"regexp"
 	"strconv"
 	"strings"
-	"time"
-
-	"regexp"
 
 	"github.com/majormjr/rcon"
 )
@@ -182,9 +179,9 @@ func initFactorio() (f *FactorioServer, err error) {
 		}
 	}
 
-    if config.autostart == "true" {
-    		go autostart()
-    }
+	if config.autostart == "true" {
+		go autostart()
+	}
 
 	return
 }
@@ -333,89 +330,6 @@ func (f *FactorioServer) writeLog(logline string) error {
 func (f *FactorioServer) checkLogError(logline []string) error {
 	// TODO Handle errors generated by running Factorio Server
 	log.Println(logline)
-
-	return nil
-}
-
-func (f *FactorioServer) Stop() error {
-	if runtime.GOOS == "windows" {
-
-		// Disable our own handling of CTRL+C, so we don't close when we send it to the console.
-		setCtrlHandlingIsDisabledForThisProcess(true)
-
-		// Send CTRL+C to all processes attached to the console (ourself, and the factorio server instance)
-		sendCtrlCToPid(0)
-		log.Println("Sent SIGINT to Factorio process. Factorio shutting down...")
-
-		// Somehow, the Factorio devs managed to code the game to react appropriately to CTRL+C, including
-		// saving the game, but not actually exit. So, we still have to manually kill the process, and
-		// for extra fun, there's no way to know when the server save has actually completed (unless we want
-		// to inject filesystem logic into what should be a process-level Stop() routine), so our best option
-		// is to just wait an arbitrary amount of time and hope that the save is successful in that time.
-		time.Sleep(2 * time.Second)
-		f.Cmd.Process.Signal(os.Kill)
-
-		// Re-enable handling of CTRL+C after we're sure that the factrio server is shut down.
-		setCtrlHandlingIsDisabledForThisProcess(false)
-
-		f.Running = false
-		return nil
-	}
-
-	err := f.Cmd.Process.Signal(os.Interrupt)
-	if err != nil {
-		if err.Error() == "os: process already finished" {
-			f.Running = false
-			return err
-		}
-		log.Printf("Error sending SIGINT to Factorio process: %s", err)
-		return err
-	}
-	f.Running = false
-	log.Printf("Sent SIGINT to Factorio process. Factorio shutting down...")
-
-	err = f.Rcon.Close()
-	if err != nil {
-		log.Printf("Error close rcon connection: %s", err)
-	}
-
-	return nil
-}
-
-func (f *FactorioServer) Kill() error {
-	if runtime.GOOS == "windows" {
-
-		err := f.Cmd.Process.Signal(os.Kill)
-		if err != nil {
-			if err.Error() == "os: process already finished" {
-				f.Running = false
-				return err
-			}
-			log.Printf("Error sending SIGKILL to Factorio process: %s", err)
-			return err
-		}
-		f.Running = false
-		log.Println("Sent SIGKILL to Factorio process. Factorio forced to exit.")
-
-		return nil
-	}
-
-	err := f.Cmd.Process.Signal(os.Kill)
-	if err != nil {
-		if err.Error() == "os: process already finished" {
-			f.Running = false
-			return err
-		}
-		log.Printf("Error sending SIGKILL to Factorio process: %s", err)
-		return err
-	}
-	f.Running = false
-	log.Printf("Sent SIGKILL to Factorio process. Factorio forced to exit.")
-
-	err = f.Rcon.Close()
-	if err != nil {
-		log.Printf("Error close rcon connection: %s", err)
-	}
 
 	return nil
 }
