+++ src/factorio_server_windows.go
@@ -39,3 +39,42 @@ func setCtrlHandlingIsDisabledForThisProcess(disabled bool) {
 		log.Fatalf("SetConsoleCtrlHandler: %v\n", e)
 	}
 }
+
+func (f *FactorioServer) Kill() error {
+	err := f.Cmd.Process.Signal(os.Kill)
+	if err != nil {
+		if err.Error() == "os: process already finished" {
+			f.Running = false
+			return err
+		}
+		log.Printf("Error sending SIGKILL to Factorio process: %s", err)
+		return err
+	}
+	f.Running = false
+	log.Println("Sent SIGKILL to Factorio process. Factorio forced to exit.")
+
+	return nil
+}
+
+func (f *FactorioServer) Stop() error {
+	// Disable our own handling of CTRL+C, so we don't close when we send it to the console.
+	setCtrlHandlingIsDisabledForThisProcess(true)
+
+	// Send CTRL+C to all processes attached to the console (ourself, and the factorio server instance)
+	sendCtrlCToPid(0)
+	log.Println("Sent SIGINT to Factorio process. Factorio shutting down...")
+
+	// Somehow, the Factorio devs managed to code the game to react appropriately to CTRL+C, including
+	// saving the game, but not actually exit. So, we still have to manually kill the process, and
+	// for extra fun, there's no way to know when the server save has actually completed (unless we want
+	// to inject filesystem logic into what should be a process-level Stop() routine), so our best option
+	// is to just wait an arbitrary amount of time and hope that the save is successful in that time.
+	time.Sleep(2 * time.Second)
+	f.Cmd.Process.Signal(os.Kill)
+
+	// Re-enable handling of CTRL+C after we're sure that the factrio server is shut down.
+	setCtrlHandlingIsDisabledForThisProcess(false)
+
+	f.Running = false
+	return nil
+}
